ocr_api/pyproject.toml
[project]
name = "ocr-api"
version = "1.0.0"
description = "Microservicio FastAPI para procesamiento OCR de tickets"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pillow>=11.0.0",
    "pytesseract>=0.3.13",
    "httpx>=0.27.0",
    "result>=0.17.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
